Add scraper for Resistance Calendar
