+++ Gemfile
@@ -28,6 +28,7 @@ gem 'devise', '4.3.0'
 
 gem 'nokogiri', '1.8.0'
 gem 'httparty', '0.15.5'
+gem 'hyperclient', '0.8.6'
 gem 'cta_aggregator_client', '0.4.0'
 gem 'Indirizzo', '0.1.7'
 gem 'will_paginate', '3.1.6'
