+++ lib/scraper.rb
@@ -1,5 +1,6 @@
 require 'scraper/emilys_list'
 require 'scraper/five_calls'
+require 'scraper/resistance_calendar'
 
 module Scraper
   class << self
@@ -10,6 +11,10 @@ module Scraper
 
     def five_calls
       FiveCalls.new(ScrapeFail)
+    end
+    
+    def resistance_calendar
+      ResistanceCalendar.new(ScrapeFail)
     end
   end
 end
